\name{aiddata-package}
\alias{aiddata-package}
\alias{aiddata}
\docType{package}
\title{
Access AidData API
}
\description{ Functions to access and download AidData through the AidData 3.0
    API. Allows the user to download a list of aid projects for a given country
    and year(s), as well as information on the geolocation of aid projects (if
    available).
}
\details{
\tabular{ll}{
Package: \tab aiddata\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-09-25\cr
License: \tab GPL-3.0\cr
}

}
\author{
Felix Haass
}
\references{
Tierney, Michael J., Daniel L. Nielson, Darren G. Hawkins, J. Timmons Roberts, Michael G. Findley, Ryan M. Powers, Bradley Parks, Sven E. Wilson, and Robert L. Hicks. 2011. More Dollars than Sense: Refining Our Knowledge of Development Finance Using AidData. World Development 39 (11): 1891-1906.

Strandow, Daniel, Michael Findley, Daniel Nielson, and Joshua Powell. 2011. The UCDP-AidData codebook on Geo-referencing Foreign Aid. Version 1.1. Uppsala Conflict Data Program. Uppsala, Sweden: Uppsala University.
}
\seealso{
    \code{\link{getAid}}
    
    \code{\link{getGIS}}
}
\keyword{ Aid }
\examples{

}
