% Generated by roxygen2 (4.0.2): do not edit by hand
\name{browse_aid}
\alias{browse_aid}
\title{Browse AidData Project Details}
\usage{
browse_aid(id = NULL)
}
\arguments{
\item{id}{The unique AidData project ID (given in the \code{project_id} field
  of other \code{aiddata} functions)}
}
\description{
Opens the AidData web page with information on the given project. For
interactive use only.
}
\details{
Quick lookup function to browse individual project IDs. Opens a
  browser window with detail informationed from on the given project on the
  AidData website. This can be useful if more information about the project
  is required, e.g. the long description which isn't downloaded by
  \code{get_aid}. Wrapper around \code{httr}'s \code{\link{BROWSE}}
  function. Only works interactively.
}
\examples{
\dontrun{

browse_aid(50225207) # Opens AidData project page an Irish education project in Uganda.
}
}
\seealso{
\code{\link{BROWSE}}
}

