% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getAid}
\alias{getAid}
\title{Download AidData Project Information}
\usage{
getAid(rec = NULL, donor = NULL, start = NULL, end = NULL,
  progressbar = TRUE)
}
\arguments{
\item{rec}{Character vector of recipient countries (ISO-2 country codes, e.g.
"AO" for Angola). Donor organizations not yet implemented.}

\item{donor}{Character vector of donor countries (ISO-2 country codes, e.g.
"US"). Donor organizations not yet implemented.}

\item{start}{First year of data.}

\item{end}{Last year of data.}

\item{progressbar}{Default TRUE. Show a progress bar?}
}
\value{
Data frame in which one row represents one project.
}
\description{
Downloads all aid projects in the AidData 3.0 database for one or several
countries, parses resulting JSON files, and returns a data frame with the aid
projects.
}
\details{
This function downloads information on all aid projects between
  \code{start} and \code{end} years, either to the recipient country
  specified in \code{rec} or from the country \code{donor}. Downloading data
  for a large number of countries over a long period of time can result in a
  large number of projects and may take long; the use of the
  \code{progressbar} option is recommended.

  Not all information available through AidData is provided; due to
  performance reasons the functions is restricted to the following variables

  \itemize{ \item{\code{project_id:} }{Unique AidData project ID}
  \item{\code{recipient: } }{Recipient name, character}
  \item{\code{recipient_iso3:} }{Recipient ISO-3 character string}
  \item{\code{donor:} }{Donor name, character string}
  \item{\code{donor_iso3:} }{Donor ISO-3 character string. \code{NA} if IGO,
  NGO, or no official counry} \item{\code{year:} }{Year of commitment}
  \item{\code{title: }}{Project title} \item{\code{name: }}{Sector name}
  \item{\code{tr_constant_value: }}{Value of transaction in constant USD2009}
  \item{\code{tr_nominal_value: }}{Value of transaction, as reported by the
  donor in the reported currency} \item{\code{sector3_code: }}{AidData's
  3-digit sector code} \item{\code{sector5_code: }}{AidData's 5-digit sector
  code} }
}
\examples{
\dontrun{
 # Download and store all aid projects in Liberia by all donors
 # between 2006 and 2007 and don't show a progress bar.
 aid_projects <- getAid(rec = "LR", start = 2006, end = 2007, progressbar = FALSE)
}
}
\references{
Tierney, Michael J., Daniel L. Nielson, Darren G. Hawkins, J. Timmons
Roberts, Michael G. Findley, Ryan M. Powers, Bradley Parks, Sven E. Wilson,
and Robert L. Hicks. 2011. More Dollars than Sense: Refining Our Knowledge of
Development Finance Using AidData. \emph{World Development} 39 (11):
1891-1906.

\url{http://aiddata.org}

\url{http://aiddata.org/user-guide}
}
\seealso{
\code{\link{getGIS}}
}

