% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{lookup.table}
\alias{lookup.table}
\title{Lookup table for countries and organizations in AidData}
\format{A dataset with 234 observations on 8 variables:
\itemize{
   \item{\code{project_id:} }{Unique AidData project ID}
   \item{\code{name: }}{Organization or country name (character string)}
   \item{\code{iso3: }}{ISO-3 character code (if available)}
   \item{\code{iso2: }}{ISO-2 character code (if available)}
   \item{\code{oecdCode: }}{OECD organizational code}
   \item{\code{type_id: }}{Unclear (not provided in the documentation)}
   \item{\code{type_name: }}{Unclear (not provided in the documentation)}
   \item{\code{parent_agency_id: }}{Unclear (not provided in the documentation)}
}}
\source{
\url{http://aiddata.org}
}
\usage{
lookup.table
}
\description{
A data frame with AidData's IDs for all 234 countries and organizations. Used
internally by \code{aiddata} functions.
}
\keyword{datasets}

