% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lookuptable.orgID}
\alias{lookuptable.orgID}
\title{Generate table with AidData destination/receiver IDs}
\usage{
lookuptable.orgID(x)
}
\description{
A data frame with 234 observations on the following 8 variables.
Used internally by the \code{getAid()} function.
}

